refactor(bench): trim workflow and fixture bloat
